Fixes the accounts service

+++ src/services/account/account.service.ts
@@ -1,13 +1,15 @@
 import { Knex } from 'knex';
 import { BaseService } from '../../base-classes';
 import { contexts } from '../../lib/contexts';
-import { ServiceOperationError } from '../../lib/errors/service-operation.error';
+import { UnexpectedError, ValidationError } from '../../lib/errors';
+import { BaseCustomError } from '../../lib/errors/base-custom-error.error';
 import { IAccount } from './types/account.type';
 import {
   createAccountDtoSchema,
   ICreateAccountDto,
-  ICreateAccountOperationResult,
+  idSchema,
   IUpdateAccountDto,
+  updateAccountDtoSchema,
 } from './types/dto.type';
 import { IAccountRepository } from './types/repository.type';
 import { IAccountService } from './types/service.type';
@@ -21,17 +23,16 @@ export class AccountService extends BaseService implements IAccountService {
   async createAccount(
     newAccount: ICreateAccountDto,
     transaction?: Knex.Transaction
-  ): Promise<ICreateAccountOperationResult> {
+  ): Promise<string> {
     // @TODO: move this validation to a api middleware. this service is suposed to receive the correct data
     const validationResult = createAccountDtoSchema.safeParse(newAccount);
 
     if (!validationResult.success) {
-      const errorInstance = new ServiceOperationError({
+      const errorInstance = new ValidationError({
         details: {
           input: newAccount,
           errors: validationResult.error.issues,
         },
-        // @NOTE: add a new context, just for the account
         context: contexts.ACCOUNT_CREATE,
       });
 
@@ -47,9 +48,14 @@ export class AccountService extends BaseService implements IAccountService {
 
       this.logger.info(`new account created: ${accountId}`);
 
-      return { accountId };
+      return accountId;
     } catch (error) {
-      const errorInstance = new ServiceOperationError({
+      if (error instanceof BaseCustomError) {
+        this.logger.error(error);
+        throw error;
+      }
+
+      const errorInstance = new UnexpectedError({
         details: {
           input: newAccount,
         },
@@ -65,10 +71,40 @@ export class AccountService extends BaseService implements IAccountService {
   }
 
   async getAccountById(accountId: string): Promise<IAccount | undefined> {
+    const isIdValid = idSchema.safeParse(accountId);
+
+    if (!isIdValid.success) {
+      const errorInstance = new ValidationError({
+        details: {
+          input: { accountId },
+          errors: isIdValid.error.issues,
+        },
+        context: contexts.ACCOUNT_GET_BY_ID,
+      });
+
+      this.logger.error(errorInstance);
+      throw errorInstance;
+    }
+
     return this.accountRepository.getAccountById(accountId);
   }
 
   async removeAccountById(accountId: string): Promise<boolean> {
+    const isIdValid = idSchema.safeParse(accountId);
+
+    if (!isIdValid.success) {
+      const errorInstance = new ValidationError({
+        details: {
+          input: { accountId },
+          errors: isIdValid.error.issues,
+        },
+        context: contexts.ACCOUNT_REMOVE_BY_ID,
+      });
+
+      this.logger.error(errorInstance);
+      throw errorInstance;
+    }
+
     return this.accountRepository.removeAccountById(accountId);
   }
 
@@ -76,6 +112,36 @@ export class AccountService extends BaseService implements IAccountService {
     accountId: string,
     updateAccountDto: IUpdateAccountDto
   ): Promise<boolean> {
+    const isIdValid = idSchema.safeParse(accountId);
+
+    if (!isIdValid.success) {
+      const errorInstance = new ValidationError({
+        details: {
+          input: { accountId },
+          errors: isIdValid.error.issues,
+        },
+        context: contexts.ACCOUNT_REMOVE_BY_ID,
+      });
+
+      this.logger.error(errorInstance);
+      throw errorInstance;
+    }
+
+    const isPayloadValid = updateAccountDtoSchema.safeParse(updateAccountDto);
+
+    if (!isPayloadValid.success) {
+      const errorInstance = new ValidationError({
+        details: {
+          input: { accountId },
+          errors: isPayloadValid.error.issues,
+        },
+        context: contexts.ACCOUNT_UPDATE_BY_ID,
+      });
+
+      this.logger.error(errorInstance);
+      throw errorInstance;
+    }
+
     return this.accountRepository.updateAccountById(
       accountId,
       updateAccountDto
